+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+- Sentry Supabase Integration ([#2913](https://github.com/getsentry/sentry-dart/pull/2913))
+  - Adds the `sentry_supabase` package to instrument supabase with Sentry breadcrumbs, traces and errors.
 - Add `message` parameter to `captureException()` ([#2882](https://github.com/getsentry/sentry-dart/pull/2882))
 - Add module in SentryStackFrame ([#2931](https://github.com/getsentry/sentry-dart/pull/2931))
   - Set `SentryOptions.includeModuleInStackTrace = true` to enable this. This may change grouping of exceptions.